+++ src/api/endpoints/register.rs
@@ -1,8 +1,7 @@
 use crate::api::schemas::register_user::RegisterUserSchema;
-use crate::application::{ApiResponse, AppState};
+use crate::application::AppState;
 use crate::domain::repositories::auth_repository::AuthRepository;
-use crate::model::api_error::ApiError;
-use crate::model::user::FilteredUser;
+use crate::model::{api_error::ApiError, api_response::ApiResponse, user::FilteredUser};
 use axum::{extract::State, Json};
 use std::sync::Arc;
